requirements.txt
absl-py==2.2.1
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.5
django-jalali==7.3.0
djangorestframework==3.16.0
drf-yasg==1.21.10
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.7
namex==0.0.8
numpy==2.2.3
opencv-python==4.11.0.86
optree==0.14.1
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
railway==0.0.4
redis==5.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
yarl==1.18.3
yt-dlp==2025.2.19
